encapsulate a reusable component:TopoComponent
